Rename spec.Data type.
