+++ spec/specification.go
@@ -1,4 +1,4 @@
-package occurrences
+package spec
 
 import (
 	"fmt"
@@ -42,17 +42,12 @@ var (
 )
 
 // Data contains occurrence information.
-type Data struct {
+type Specification struct {
 	Occurrences set.Interface
 	DaysOfWeek  set.Interface
 }
 
-// String returns a string representation of the type.
-func (t Data) String() string {
-	return fmt.Sprintf("Data{occurrences=%s, daysOfWeek=%s}", t.Occurrences, t.DaysOfWeek)
-}
-
-func (t Data) FriendlyStrings(yearMode bool) (results []string) {
+func (t Specification) FriendlyStrings(yearMode bool) (results []string) {
 	occurrences := set.IntSlice(t.Occurrences)
 	sort.Ints(occurrences)
 
@@ -73,7 +68,7 @@ func (t Data) FriendlyStrings(yearMode bool) (results []string) {
 }
 
 // Specification returns a string representation of the Occurrences structure.
-func (t Data) Specification() string {
+func (t Specification) String() string {
 	occurrences := set.IntSlice(t.Occurrences)
 	var output []string
 
@@ -106,7 +101,7 @@ func dayOfWeekSlice(s set.Interface) (results []time.Weekday) {
 }
 
 // Intersects returns true if the other Data instance intersects this instance.
-func (t Data) Intersects(other *Data) bool {
+func (t Specification) Intersects(other *Specification) bool {
 	oIntersect := !set.Intersection(t.Occurrences, other.Occurrences).IsEmpty()
 	dIntersect := !set.Intersection(t.DaysOfWeek, other.DaysOfWeek).IsEmpty()
 	return oIntersect && dIntersect
@@ -156,7 +151,7 @@ func Validate(specification string, yearMode bool) error {
 }
 
 // New creates a new occurrence data instance.
-func New(specification string) (*Data, error) {
+func New(specification string) (*Specification, error) {
 	parts := strings.Split(specification, "#")
 
 	occurrences, err := toIntSet(strings.Split(parts[0], ","))
@@ -169,13 +164,13 @@ func New(specification string) (*Data, error) {
 		return nil, err
 	}
 
-	ds := &Data{occurrences, daysOfWeek}
+	ds := &Specification{occurrences, daysOfWeek}
 	return ds, nil
 }
 
 // NewFromDate creates a new occurrence data instance
 // from the provided date.
-func NewFromDate(date time.Time, forYear bool) *Data {
+func NewFromDate(date time.Time, forYear bool) *Specification {
 	day := date.Day()
 	if forYear {
 		day = date.YearDay()
@@ -190,7 +185,7 @@ func NewFromDate(date time.Time, forYear bool) *Data {
 	occurrences.Add(int(occOrd))
 	daysOfWeek := set.New(set.ThreadSafe)
 	daysOfWeek.Add(date.Weekday())
-	return &Data{occurrences, daysOfWeek}
+	return &Specification{occurrences, daysOfWeek}
 }
 
 func toIntSet(values []string) (set.Interface, error) {
